Handle when --fastqs is specified on the CLI while CLI options are also in a sample sheet (#33)
* [bugfix] do not require FASTQs on the command line

The fastqs clap arugment had `required = True` for `fastqs`.  However,
when fastqs is provided on the command line and there are also command
line arguments in the sample sheet, a missing argument error is
encountered due to "fastqs not being provided".
This occurs when trying to use the `try_update_from` on `Opts`.  A
reproducible example independent of this code base can be found here:
https://github.com/clap-rs/clap/issues/4617

* [internal] minor doc fixup in sample metadata validate_samples

* add unit test for pre-specified FASTQs with sample sheet