[feat] add an option to only process specific lanes (#36)
* [feat] add an option to only process specific lanes

Only samples (in the sample sheet) and FASTQs (with known lane numbers)
will be processed that match the provided list of lanes.  This option
cannot be used when the lane for a FASTQ is not known (e.g. when
the FASTQS are explicitly given on the command line).

Also allow the same sample_id/barcode combination is used across lanes for a
given sample.  Previously an error would occur, now the repeated sample
is treated as one sample (just in many lanes).

Other changes:
- the read name is only checked up to the first colon within a group of
  FASTQS (e.g. R1 across lanes)
- print out the # of threads being used
- bug fix start sample ordinal at zero