[bugfix] reads were written to the wrong file when --lanes was used
When subsetting to a specific lane, the reads were being written to the
wrong output file.  This was caused by the list of samples internally
not being in ascending order by ordinal, but the writers created after
sorting the samples by ordinal.

The fix is to order the samples by ordinal prior to creating the
writers, and adding in validation code to ensure the samples are sorted
in ascending order by ordinal.

Also changed the ordinal of the undetermined sample to be one larger
than the largest ordinal in the existing samples.